+++ examples/webbilling-demo/src/pages/rc_paywall_no_offering_passed/index.tsx
@@ -11,6 +11,7 @@ const RCPaywallNoOfferingPassedPage: React.FC = () => {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const lang = searchParams.get("lang");
+  const email = searchParams.get("email");
   const {
     openSettings,
     settings: { customVariables },
@@ -29,6 +30,7 @@ const RCPaywallNoOfferingPassedPage: React.FC = () => {
         htmlTarget: document.getElementById("paywall") || undefined,
         selectedLocale: lang || undefined,
         customVariables,
+        customerEmail: email || undefined,
       })
       .then((purchaseResult: PurchaseResult) => {
         const { customerInfo, redemptionInfo } = purchaseResult;
@@ -47,7 +49,7 @@ const RCPaywallNoOfferingPassedPage: React.FC = () => {
         );
       })
       .catch((err: Error) => console.log(`Error: ${err}`));
-  }, [navigate, lang, customVariables]);
+  }, [navigate, lang, customVariables, email]);
 
   return (
     <>