+++ src/ui/express-purchase-button/express-purchase-button.svelte
@@ -362,7 +362,6 @@
       onReady={onExpressCheckoutElementReady}
       {expressCheckoutOptions}
       forceEnableWalletMethods={false}
-      billingAddressRequired={brandingInfo?.gateway_tax_collection_enabled}
       hideCheckoutSeparator={true}
     />
   {/if}